Added developer settings
Hidden from release builds
