+++ app/build.gradle
@@ -44,12 +44,9 @@ dependencies {
     implementation 'androidx.constraintlayout:constraintlayout:2.0.0-beta2'
     implementation 'de.hdodenhof:circleimageview:3.0.1'
     implementation 'com.github.yalantis:ucrop:2.2.4-native'
-    implementation 'ca.rmen:lib-sunrise-sunset:1.1.1'
-    implementation 'com.google.android.gms:play-services-location:17.0.0'
+    implementation "androidx.preference:preference:1.1.0"
     testImplementation 'junit:junit:4.12'
 }
-
-
 
 
 apply plugin: 'com.google.gms.google-services'
