+++ app/src/main/java/com/marco97pa/puntiburraco/SettingActivity.java
@@ -82,7 +82,8 @@ public class SettingActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
 
         // Display the fragment as the main content.
-        getFragmentManager().beginTransaction()
+        getSupportFragmentManager()
+                .beginTransaction()
                 .replace(android.R.id.content, new SettingsFragment())
                 .commit();
