+++ app/src/main/res/xml/method_preference.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.preference.PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+    <ListPreference
+        android:key="insert_method"
+        android:title="Insert Method"
+        android:defaultValue="3"
+        android:entries="@array/input_methods"
+        android:entryValues="@array/input_methods_values"
+        android:summary="Select a method"
+        android:icon="@drawable/ic_arrow_back_black_24dp"
+        />
+</androidx.preference.PreferenceScreen>
